Create AVG Connected Sessions in a Custom Group
**Requires the 'vRealize Operations Management Pack for Horizon' to work**

This formula will show the average number of connected sessions with a custom group.

You'll need to first have or create a custom group which organizes the different VDI pools. For example, you might want to create a custom group for
the VDI pools based on business unit, a specific user group, or maybe by location. It all depends upon how your company I is setup and you are using your different VDI pools.

When you create the super metric, you'll need to associate the object type to it.